Refactored to adhere to PR comments
