+++ backend/app/environment_platform_configurations/schemas/__init__.py
@@ -1,0 +1,5 @@
+from .aws_schema import AWSEnvironmentPlatformConfiguration
+
+__all__ = [
+    "AWSEnvironmentPlatformConfiguration",
+]
