+++ backend/app/environment_platform_configurations/schemas/aws_schema.py
