+++ backend/app/environment_platform_service_configurations/schemas/__init__.py
@@ -1,0 +1,4 @@
+from .glue_schema import AWSGlueConfig
+from .s3_schema import AWSS3Config
+
+__all__ = ["AWSGlueConfig", "AWSS3Config"]
