+++ backend/app/environment_platform_service_configurations/schemas/config_schema.py
