+++ backend/app/environment_platform_service_configurations/schemas/s3_schema.py
@@ -1,6 +1,4 @@
-from app.environment_platform_service_configurations.config_schema import (
-    BaseEnvironmentPlatformServiceConfigurationDetail,
-)
+from .config_schema import BaseEnvironmentPlatformServiceConfigurationDetail
 
 
 class AWSS3Config(BaseEnvironmentPlatformServiceConfigurationDetail):
