+++ backend/app/environments/schema.py
@@ -3,11 +3,8 @@ from uuid import UUID
 from app.shared.schema import ORMModel
 
 
-class EnvironmentCreate(ORMModel):
+class Environment(ORMModel):
+    id: UUID
     name: str
     context: str
     is_default: bool = False
-
-
-class Environment(EnvironmentCreate):
-    id: UUID
