+++ backend/app/platform_services/model.py
@@ -6,13 +6,12 @@ from sqlalchemy.dialects.postgresql import UUID
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
 from app.database.database import Base
-from app.shared.model import BaseORM
 
 if TYPE_CHECKING:
     from app.platforms.model import Platform
 
 
-class PlatformService(Base, BaseORM):
+class PlatformService(Base):
     __tablename__ = "platform_services"
 
     id = Column(UUID(as_uuid=True), primary_key=True, default=uuid.uuid4)
