+++ backend/app/platforms/schema.py
@@ -3,9 +3,6 @@ from uuid import UUID
 from app.shared.schema import ORMModel
 
 
-class PlatformCreate(ORMModel):
-    name: str
-
-
-class Platform(PlatformCreate):
+class Platform(ORMModel):
     id: UUID
+    name: str
